CHECKPOINT : TOUT MARCHE !!!

+++ src/utils/restaurantStarsAverage.js
@@ -2,7 +2,7 @@ export const restaurantStarsAverage = (arr) => {
       let arrLength = arr.length, starsSum = 0, i;
 
       for (i = 0; i < arrLength; i++) {
-            starsSum += parseInt(arr[i]);
+            starsSum += parseFloat(arr[i]);
       }
 
       return (starsSum/arrLength).toFixed(2);
